Switched to Nintendump.py
Nintendump.py is a rewrite of dumpinfo.idc to solve some issues I was
having with with quotations in comments. I didn't feel like learning how
to use idc (the documentation is terrible) so I just rewrote the entire
thing in Python.

Nintendump also supports exporting the functions as a *.map file to use
within Dolphin's debugger. :)